+++ Source/Domain/Model/Game/Size.cs
@@ -1,6 +1,6 @@
 ﻿namespace VttTools.Model.Game;
 
 public class Size {
-    public int Width { get; set; }
-    public int Height { get; set; }
+    public double Width { get; set; }
+    public double Height { get; set; }
 }
